Rename and move file
- Rename Bookmarks to BookmarkPage
- Move BookmarkPage to BookmarkPage folder
- Update the path in app.jsx
